+++ apps/classroomio-com/src/lib/Home/Hero.svelte
