+++ apps/classroomio-com/src/routes/+page.svelte
@@ -1,20 +1,19 @@
 <script>
-  import Hero from '$lib/Hero/Hero.svelte';
-  import Body from '$lib/Body/Body.svelte';
-  import Faq from '$lib/Faq/Faq.svelte';
-  import StudentDashboard from '$lib/StudentDashboard/StudentDashboard.svelte';
-  import MoreFeatures from '$lib/MoreFeatures/MoreFeatures.svelte';
-  import Testimonial from '$lib/Testimonial/Testimonial.svelte';
+  import Hero from '$lib/Home/Hero.svelte';
+  import CoreFeatures from '$lib/Home/CoreFeatures.svelte';
+  import Faq from '$lib/Home/Faq.svelte';
+  import StudentDashboard from '$lib/Home/StudentDashboard.svelte';
+  import MoreFeatures from '$lib/Home/MoreFeatures.svelte';
+  import Testimonial from '$lib/Home/Testimonial.svelte';
 
   import { PUBLIC_ENABLE_FAQ } from '$env/static/public';
 </script>
 
 <Hero />
-<Body />
+<CoreFeatures />
 <StudentDashboard />
 <Testimonial />
 <MoreFeatures />
-
 
 {#if PUBLIC_ENABLE_FAQ}
   <Faq />
